Adding Page Name configuration to Reports.
When adding a Power BI `Report` element, the `Page Name` element is not
present. If the report that is being embedded has more than one page,
that means only the report's default page will be visible.

This adds the `Page Name` setting to `Report` items, such that the user
can now configure which page from the report should be displayed.

If no value is provided, or if a non-existent identifier is given, the
report's default page is shown.